requirements.txt
pyaudio == 0.2.14
librosa == 0.10.1
sounddevice == 0.4.6
numpy == 1.26.4
pillow == 10.3.0
SpeechRecognition == 3.10.3